Removal of empty method
There appeared to be a method found within the GZIPResponseWrapper.java file that was missing a method body. There are no variables declared and called upon making this an invalid method, it simply becomes omitted they may cause errors or unusual behavior within the code. 
I have simply removed the setContentLength method and have left a comment as to why that is indeed incorrect and why it was removed